% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_div.R
\name{bdiv.f}
\alias{bdiv.f}
\title{A function to derive LCBD and SCBD from function beta.div}
\usage{
bdiv.f(df, ...)
}
\arguments{
\item{df}{a dataframe containing species abundance/incidence with rows as sites and columns as species}

\item{...}{further arguments for function beta.div}
}
\value{
a list containing two dataframes LCBD, SCBD
}
\description{
A function to derive LCBD and SCBD from function beta.div
}
