% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_mean.sd.R
\name{cal_mean.sd}
\alias{cal_mean.sd}
\title{Title Calculated mean±sd}
\usage{
cal_mean.sd(df, digits = NULL, nsmall = NULL)
}
\arguments{
\item{df}{a numeric vector}

\item{digits}{how many digits}

\item{nsmall}{the minimum number of digits to the right of the decimal point in formatting real/complex numbers in non-scientific formats. Allowed values are 0 <= nsmall <= 20.}
}
\value{
a character
}
\description{
Title Calculated mean±sd
}
\examples{
usuallt it is used with dplyr::summarise and dplyr::mutate funciton
}
