% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_taxo.div.R
\name{cal_taxo.div}
\alias{cal_taxo.div}
\title{Title Calculte taxonomic diversity}
\usage{
cal_taxo.div(spe.df)
}
\arguments{
\item{spe.df}{a dataframe, row as site while column as species name, nonumeric data will be used in final results}
}
\value{
a dataframe, Species number, Shannon-wiener diversity, Simpson diversity and Evenness are calculated
}
\description{
Title Calculte taxonomic diversity
}
