% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_fourmula.R
\name{construct_fourmula}
\alias{construct_fourmula}
\title{Title Construct formula with multipe response and explanatory variabels}
\usage{
construct_fourmula(Resp.Var, Exp.Vars, type = NULL, collapse = NULL)
}
\arguments{
\item{Resp.Var}{response variables, should be character vectors}

\item{Exp.Vars}{explanatory variables, should be character vectors}

\item{type}{construct formula for each Resp.Vars and all each Exp.Vars (default) or
for each Resp.Vars and each Exp.Vars(set an arbitary value to type)}

\item{collapse}{build addititive (+) models or models with interactions (*)}
}
\value{

}
\description{
Title Construct formula with multipe response and explanatory variabels
}
