% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2df.R
\name{dist2df}
\alias{dist2df}
\title{Title Convert dist and matrix type data into dataframe}
\usage{
dist2df(inDist)
}
\arguments{
\item{inDist}{a dist or matrix type data}
}
\value{
a dataframe
}
\description{
Title Convert dist and matrix type data into dataframe
}
