% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envls_trans.R
\name{envls_trans}
\alias{envls_trans}
\title{Transformation for a list with environmental variables in dataframe}
\usage{
envls_trans(env_ls, env.trans = NULL)
}
\arguments{
\item{env_ls}{a list with environmental variables in dataframe}

\item{env.trans}{methods for trasnformation}
}
\value{
a list
}
\description{
Transformation for a list with environmental variables in dataframe
}
