% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.mntd.R
\name{fst.comdistnt}
\alias{fst.comdistnt}
\title{fast computation of inter-community mean nearest taxon distance (bNTI)}
\usage{
fst.comdistnt(
  comm,
  phy_dist,
  exclude_conspecifics = FALSE,
  nworkers = NULL,
  ses = F,
  abundance_weighted = FALSE,
  runs = NULL
)
}
\arguments{
\item{comm}{a data.frame containing species abundance}

\item{phy_dist}{a matrix containing phylogenetic distance}

\item{exclude_conspecifics}{Should conspecific taxa in different communities be exclude from MNTD calculations? (default = FALSE)}

\item{nworkers}{numbers of cores used for computation}

\item{ses}{if ses.bNTI or NTI be calculated}

\item{abundance_weighted}{whether bNTI is abundance weighted}

\item{runs}{999 random draws, can be set to to smaller number to reduce computation time}
}
\description{
fast computation of inter-community mean nearest taxon distance (bNTI)
}
