% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.mntd.R
\name{fst.ses.mpd}
\alias{fst.ses.mpd}
\title{Fast computation of MPD and SES.MPD}
\usage{
fst.ses.mpd(
  comm,
  phy.dist,
  null_model = "taxaShuffle",
  nworkers = NULL,
  ses = F,
  abundance_weighted = F,
  runs = NULL
)
}
\arguments{
\item{comm}{a data.frame containing species abundance}

\item{phy.dist}{a matrix containing phylogenetic distance}

\item{null_model}{only implemented for random taxa.labels}

\item{nworkers}{numbers of cores used for computation}

\item{ses}{if SES.MPD should be calculated}

\item{abundance_weighted}{whether MPD/SES.MPD is abundance weighted}

\item{runs}{usually 999 random draws, can be set to to smaller number to reduce computation time}
}
\description{
Fast computation of MPD and SES.MPD
}
