% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.mntd.R
\name{fst.ses.pd}
\alias{fst.ses.pd}
\title{fst.ses.pd a function for calculating PD}
\usage{
fst.ses.pd(
  comm,
  phy,
  null_model = "taxaShuffle",
  nworkers = NULL,
  ses = NULL,
  runs = NULL
)
}
\arguments{
\item{comm}{a matrix contains site * species information, note that it must be a matrix
(will be automatically converted to sparse matrix) or sparse matrix.}

\item{phy}{a phylogenetic (or generated from functional distance) tree}

\item{null_model}{only implemented for taxaShuffle}

\item{nworkers}{numbers of cores for calculating, by default it only 1 core will be used.}

\item{ses}{a Boolean value indicates whether standardized form of PD will be calculated}

\item{runs}{an integer indicates numbers of randomization, by default 999, can be set to smaller values for tesing.}
}
\value{
a tibble contains observed PD and standardized PD (when ses = T)
}
\description{
fst.ses.pd a function for calculating PD
}
