% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aov.letter.R
\name{get_aov.letter}
\alias{get_aov.letter}
\title{Title Get multiple comparison letters from anova (HSD.test,package agricolae )}
\usage{
get_aov.letter(resp.var, exp.var, data, p.adj = NULL)
}
\arguments{
\item{resp.var}{a character regarding the numeric data}

\item{exp.var}{a character regarding the group factor}

\item{data}{a dataframe}

\item{p.adj}{methods for adjuted p value, see function aruguments methods in function p.adjust for more infomation}
}
\value{

}
\description{
Title Get multiple comparison letters from anova (HSD.test,package agricolae )
}
