% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr_df.R
\name{get_corr_df}
\alias{get_corr_df}
\title{Title A function return correlation and p value in dataframe formate}
\usage{
get_corr_df(df.x, df.y = NULL, type = NULL)
}
\arguments{
\item{df.x}{a dataframe, only numeric data will be used}

\item{df.y}{a dataframe, not necessary}

\item{type}{pearson (by defacult) or spearman correlation to be used}
}
\value{

}
\description{
Title A function return correlation and p value in dataframe formate
}
