% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files_dir.R
\name{get_files_dir}
\alias{get_files_dir}
\title{A function that return list of file naes with specified patterns}
\usage{
get_files_dir(directory = NULL, pattern = NULL)
}
\arguments{
\item{directory}{a string, absolute file path}

\item{pattern}{specified patterns e.g., "txt", "xlsx", and etc.}
}
\value{
a list, file path as the element while file name as the element name
}
\description{
A function that return list of file naes with specified patterns
}
