% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ind_val.R
\name{get_ind_val}
\alias{get_ind_val}
\title{Conduct indicator species analsyis}
\usage{
get_ind_val(spe, grp, numitr = NULL, p.cut = NULL, p.adj = NULL, ...)
}
\arguments{
\item{spe}{a data frame wirh row as observations (sites) while column as species}

\item{grp}{a vector indicating the group of, it should be a factor}

\item{numitr}{number of iteration,deafult numitr=999}

\item{p.cut}{significant levels, deafult p.cut=0.05}

\item{p.adj}{indicate whethe adjusted p value shoul be used, by default T}

\item{...}{other param for function indval}
}
\value{
a dataframe with 5 columns i.e., species names (spe), group (group),
indicator values (indval), pvalue, freq
}
\description{
Conduct indicator species analsyis
}
