% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lm_summary.R
\name{get_lm_summary}
\alias{get_lm_summary}
\title{Get tidy results from linear, nls, lmer, lme mode}
\usage{
get_lm_summary(lm.mod, CI = NULL, k = NULL)
}
\arguments{
\item{lm.mod}{a model returned by function lm, nls, lmer, lme}

\item{CI}{whether return confidence interval for the estimates, only works for lm model}

\item{k}{numbers of digits}
}
\value{

}
\description{
Get tidy results from linear, nls, lmer, lme mode
}
