% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpa_df.R
\name{get_vpa_df}
\alias{get_vpa_df}
\title{Get tidy data from results generated by my_vpa}
\usage{
get_vpa_df(vpa.res)
}
\arguments{
\item{vpa.res}{results generated by function my_vpa}
}
\value{
a list contains 1) mod.sel, dataframe with selected variables from function adespatial:foward.sel
2) envfrac.df, explained variation by different environmental matrices
}
\description{
Get tidy data from results generated by my_vpa
}
