% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_otu.R
\name{group_otu}
\alias{group_otu}
\title{Title Group taxa (mainly for OTU) into dominant and rare}
\usage{
group_otu(df, lower = NULL, upper = NULL, simplify = T)
}
\arguments{
\item{df}{species abundance data, see https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-019-0749-8#Sec28
for how the grouping things works}

\item{lower}{0.0001}

\item{upper}{0.01}

\item{simplify}{by default simpplify=F, which will return AAT, CAT, CRAT, MT, CRT, ART; otherwise only dominant and rare taxa will return}
}
\value{

}
\description{
Title Group taxa (mainly for OTU) into dominant and rare
}
