% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_taxa.R
\name{group_taxa}
\alias{group_taxa}
\title{Group taxa  into dominant and rare sub-communities}
\usage{
group_taxa(df, method, lower = NULL, upper = NULL, simplify = T)
}
\arguments{
\item{df}{}

\item{lower}{0.0001}

\item{upper}{0.01}

\item{simplify}{by default simplify=F, which will return AAT, CAT, CRAT, MT, CRT, ART; otherwise only dominant and rare taxa will return}
}
\description{
Group taxa  into dominant and rare sub-communities
}
