% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediateMessage.R
\name{immediateMessage}
\alias{immediateMessage}
\title{Diagnostic messages when using furrr for parellization}
\usage{
immediateMessage(..., domain = NULL, appendLF = TRUE)
}
\arguments{
\item{appendLF}{}
}
\description{
Diagnostic messages when using furrr for parellization
}
