% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.mntd.R
\name{md2nat}
\alias{md2nat}
\title{A function to calculate MPD or MNTD between native and aliens species}
\usage{
md2nat(
  comm,
  phy.dist,
  native_sp.ls,
  invasive_sp.ls,
  ses = F,
  method = "mpd",
  abundance_weighted = F,
  nworkers = NULL,
  runs = NULL
)
}
\arguments{
\item{comm}{a data.frame containing species abundance}

\item{phy.dist}{a matrix containing phylogenetic distance}

\item{native_sp.ls}{a vector containing native species}

\item{invasive_sp.ls}{a vector containing native species}

\item{ses}{if ses.bNTI or bNTI be calculated}

\item{method}{MPD (default) or MNTD}

\item{abundance_weighted}{should be indcies be abundance-weighted (default FALSE)}

\item{nworkers}{numbers of cores used for computation}

\item{runs}{999 random draws, can be set to to smaller number to reduce computation time}
}
\description{
A function to calculate MPD or MNTD between native and aliens species
}
