% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_log1p.R
\name{my_log1p}
\alias{my_log1p}
\title{Title Customed log1p for data with negative value less than -1}
\usage{
my_log1p(x, base = exp(1))
}
\arguments{
\item{x}{a numeric vector}

\item{base}{exp(1) or 10, or 2 based log (x+a)}
}
\value{
a dataframe
}
\description{
Title Customed log1p for data with negative value less than -1
}
