% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ordination.R
\name{my_ordination}
\alias{my_ordination}
\title{Customized function for PCoA and NMDS}
\usage{
my_ordination(sp.df, sp.dist = NULL, method = NULL, axes = NULL, ...)
}
\arguments{
\item{sp.df}{Species abundance data, row as sites while column as species}

\item{sp.dist}{WHich distance should be used, by default, bray distance will be used}

\item{method}{Either PCoA or NMDS should be conducted, by default, PCoA will be performed with ape::pcoa.
Since it is recommended that PCoA are suitable for most of the applications.}

\item{axes}{Indicating the numbers of axes interested, by default axes=2.}

\item{...}{More arguments can be passed to vegan::metaMDS or ape::pcoa}
}
\value{
A list including 1) the ordination model and 2) the summary for further visualizations
}
\description{
Customized function for PCoA and NMDS
}
