% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_vpa.R
\name{my_vpa}
\alias{my_vpa}
\title{Conduct VPA after CCA or RDA}
\usage{
my_vpa(
  sp,
  env_ls,
  sp.trans = NULL,
  env.trans = NULL,
  mode = NULL,
  force.mode = NULL,
  ...
)
}
\arguments{
\item{sp}{a dataframe, row as sites while column as species names}

\item{env_ls}{a list, with environmental variables data as elements}

\item{sp.trans}{a string, method for species data transformation}

\item{env.trans}{a string, method for envrionmental data transformation}

\item{mode}{a string, when force.mode=F (default), it indicates which model to use when 3<dca.axlength<4
when force.mode=T, it indicates which mode to conduct forcely}

\item{force.mode}{T or F (default)}

\item{...}{further argumens for adespatial::forward.sel}
}
\value{
a list
}
\description{
Conduct VPA after CCA or RDA
}
