% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_partitioning.R
\name{rand_var.part}
\alias{rand_var.part}
\title{extract explained variation for each nested level in lmer model}
\usage{
rand_var.part(mod)
}
\arguments{
\item{mod}{a model results returned by **lme4::lmer**}
}
\value{
a dataframe contains explained variation for each level
}
\description{
extract explained variation for each nested level in lmer model
}
\details{
see https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/1365-2435.13723 for details
"We applied a linear mixed model to partition the variance in root traits with a given trait as the
dependent factor and only random effects (‘elevation’ (i.e. conspecific plants between elevations) nested in ‘species’ nested in ‘growth form’].
The remaining variance was explained by trait differences between conspecific plants growing
in different replicate plots at the same elevation (Albert et al., 2010)."

note that the random terms are ranked in this way: (1|Level3/Level2/Level1). Additionally, the order of these
terms do not affect the results.
}
