% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveall saveall.R
\name{saveall}
\alias{saveall}
\title{Title Save a list with non-dataframe data into csv/txt file}
\usage{
saveall(list_data, data.name, ...)
}
\arguments{
\item{list_data}{a list}

\item{data.name}{file name for saving the list e.g., "all.csv"}

\item{...}{further arguments for write.table}
}
\value{

}
\description{
Title Save a list with non-dataframe data into csv/txt file
}
