% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbased_vpa.R
\name{spbased_vpa}
\alias{spbased_vpa}
\title{Conduct species based VPA}
\usage{
spbased_vpa(
  sp,
  env_ls,
  sp.trans = NULL,
  env.trans = NULL,
  mode = "rda",
  force.mode = T,
  ...
)
}
\arguments{
\item{sp}{a dataframe, row as sites while column as species names}

\item{env_ls}{a list, with environmental variables data as elements}

\item{sp.trans}{a string, method for species data transformation}

\item{env.trans}{a string, method for envrionmental data transformation}

\item{mode}{when species based VPA are conducted mode="rda"}

\item{force.mode}{force.mode=T, when species based VPAs are conducted}

\item{...}{}
}
\description{
Conduct species based VPA
}
