% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spM_trans.R
\name{spM_trans}
\alias{spM_trans}
\title{Species matrix transformation}
\usage{
spM_trans(sp, sp.trans = NULL)
}
\arguments{
\item{sp}{species abundance data, row as sites while column as species names.
Non-numeric data are allowed but will be droped in the final results}

\item{sp.trans}{methods for tranformation, allowed methods are: methods in vegan::decostand, log1p, log and
other self composed functions.}
}
\value{
a dataframe
}
\description{
Species matrix transformation
}
