% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.mntd.R
\name{fst.ses.mntd}
\alias{fst.ses.mntd}
\title{Title fast computation of MNTD/SES.MNTD}
\usage{
fst.ses.mntd(
  comm,
  phy.dist,
  null_model = "taxaShuffle",
  nworkers = NULL,
  ses_mntd = NULL,
  abundance_weighted = F,
  runs = NULL
)
}
\arguments{
\item{comm}{a data.frame containing species abundance}

\item{phy.dist}{a matrix containing phylogenetic distance}

\item{null_model}{only implemented for random taxa.labels}

\item{nworkers}{numbers of cores used for computation}

\item{ses_mntd}{if SES.MPD should be calculated}

\item{abundance_weighted}{whether MNTD/SES.MNTD is abundance weighted}

\item{runs}{usually 999 random draws, can be set to to smaller number to reduce computation time}
}
\description{
Title fast computation of MNTD/SES.MNTD
}
