% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_vpa.R
\name{my_vpa}
\alias{my_vpa}
\title{Title  community- and species- based variation partitioning}
\usage{
my_vpa(
  sp,
  env_ls,
  sp.trans = NULL,
  env.trans = NULL,
  mode = NULL,
  sp.based = NULL,
  ...
)
}
\arguments{
\item{sp}{species abundance data with sites as row while species as columns, raw data is needed if you want to run species-based VPA}

\item{env_ls}{a list containing at least two elements, while at most four elements;
meaningful name are highly recommended, otherwise env1,env2 and etc will be used as the name}

\item{sp.trans}{transformation for spe data}

\item{env.trans}{transformation methos for env data, can be a vector with length 1 or with same length with env-ls}

\item{mode}{which mode to use i.e., RDA or CCA, while do community- based VPA,
the mode is chosed based on the max value of axis length in DCA analysis, i.e., axis length>4, CCA;
3<= axis length <= 4, either RDA or CCA is fine, by default, here use RDA; axis length<3, RDA}

\item{sp.based}{species-based VPA or not, by default community-based VPA will be conducted}

\item{...}{arguments for forward.sel in package adespatial}
}
\value{
a list including raw data, model information, model results
}
\description{
Title  community- and species- based variation partitioning
}
