% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_decimal.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Specify decimal number for numeric data}
\usage{
specify_decimal(x, k = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{k}{numbers of digits}
}
\value{
a character
}
\description{
Specify decimal number for numeric data
}
